+++ src/pages/user-pages/home/HomePage.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import './home-page.scss';
 
-import { homePageDataFetch } from "../../../services/data-fetch/user-pages-data-fetch/home-page/homePageDataFetch";
+import { homePageDataFetch } from "../../../services/data-fetch/user-pages-data-fetch/home-page-data-fetch/homePageDataFetch";
 
 export const HomePage = () => {
   const [headerInfo, setHeaderInfo] = useState("");
